Adding a `div`
A `div` is a generic container in HTML.  Without styling, it inherits
the position and size of the containing (or parent) HTML element—in
this case, `body`.  Go ahead and open up the file in your browser
(we’ll be using Chrome for these tutorials) and right click somewhere
on the page and select “Inspect Element.”  A window should show up
which lets you mouse over the different elements in the barebones HTML
page and see things like their dimensions and styling.