Giving our div a height, width, and color within the HTML page
While you can write a separate CSS file which will style the appearance
of your HTML elements (and indeed, we’ll do so later), you can also
write CSS directly within an element in your HTML.  This is what we’ve
done here—`style` is an attribute of our div.  Note that this attribute
is set with a sequence of key/value pairs (e.g. ‘width’ is a key and
‘100px’ is its value), separated by semicolons.

Reload the page to see the effect.