Adjusting the position of our blue box
There are a variety of ways to control the position of an HTML element.
By setting the `position` attribute to `fixed`, we’re telling the
browser that we want to define the position of top and left sides of
this `div` _relative to the position of our browser window_.  So you’ll
notice that now, when you reload the page, the square appears in the
very top and left corner of your screen, whereas before there was a
small margin.  (This small margin is a default of the `body` tag; by
setting the position attribute explicitly we are overriding this.)

Go ahead and open up the inspector and fiddle with the position
attributes to see what happens.  Experiment with values like `50%` and
`2em`—we’ll talk more about these later.