Pulling our CSS into an external stylesheet
Not only would it become unwieldy if we specified all of our styles
inline, but much of the power of CSS comes from being able to write
rules which apply to different elements on an HTML page which may share
different properties (tag name, class, parent element, etc.).

To accommodate this, it is common to put your CSS in a separate file
where these rules are defined.  But we now need to be able to tell the
browser where to look to load the CSS—this is what the `link` tag in
the `head` is for.  The new `style.css` file simply contains the CSS
rules we’d inserted inline, but in a form which causes the rules to be
applied to all `div` tags on pages loading this `stye.css` file.  The
`div` in the CSS file is known as a ‘selector’ since it selects those
elements to which the enclosed CSS rules apply.  We’ll explore later
the different ways of targeting tags with a variety of selectors.