Introducing pseudo selectors (handling events in CSS)
In CSS there are a variety of events which can trigger the application
of different CSS rules.  In this example, we’re defining a rule which
sets the background color of an element when you hover over it.  The
reason that these are called ‘pseudo’ selectors is a little pedantic:
because there is no element in the HTML which contains `hover` there is
no tag which is ‘selected’ per se.  Instead, based on an interaction
with it, the browser triggers the application of a CSS rule.  In this
case, we are setting the `background-color` attribute to green (CSS
knows [147 colors by
name](http://www.colors.commutercreative.com/grid/), we’ll talk later
about other ways of specifying colors more finely).