Introducing CSS transforms
As of CSS3, we’re now able to apply a variety of geometric transforms
(e.g. scale, rotate, translate) to our HTML elements. You can use this
to both statically adjust the appearance of your elements as well as
animate such transitions. In this example, we’ve applied a transform
to rotate our box 45 degrees (by default, clockwise). Notice that this
rotation is effected with respect to the center of the HTML element.

Also notice that all of the transitions being applied are applied
simultaneously; our `transition` rule means that the rotation and the
background color change happen in parallel, taking 1000ms, with the
ease function.