Tweaking positioning and nested divs
Here we take one of our divs and insert it into another, making `one`
the child of `two`.  worth noting is that we’ve defined a new, more
specific rule where our rules by `id` are overriding our general `div`
styling.  We also explore different ways of defining positioning rules
relative to the window, parents, and so on to take a look at how the
CSS position attributes of `relative`, `absolute`, `static`, and
`fixed` work.

This should help you to understand how the basics of [the Portfolio
project](https://github.com/DGMD-E-15/Portfolio/) are put together,
with the `nav` tag playing the role of `#two` here and the individual
projects like `#one` here.