Re-constructing our page via JavaScript
Just as with CSS, you can include JavaScript inline or in an external
file.  One of the things you can do with JavaScript is manipulate
HTML—create new nodes, add attributes to them, etc.  This is often
referred to as “manipulating the Document Object Model (DOM)” (which
you can read more about
[here](https://stackoverflow.com/questions/14094697/javascript-how-to-cr
eate-new-div-dynamically-change-it-move-it-modify-it-in)).

We’ve left the original HTML we’re reconstructing in as a comment for
reference.