Animating motion of the square using JavaScript and a time via setInterval
Here we set up a timer to move the square one pixel down and to the
right every 50 milliseconds.  We also write a basic utility function to
make it easier to grab the current position of the square.

`moveSquare` adjusts the `top` and `left` attributes to simulate
motion.  `getPosition` makes it easy for us to grab the current
position attributes of an element and hiding some subtleties of CSS
rendering which we’ll get to later.

We’re using
[`requestAnimationFrame`](http://creativejs.com/resources/requestanimati
onframe/) to address some of the same underlying issues as came up in
the original Portfolio’s scrolling jitter—